my_maf/requirements.txt
asyncio
requests
agent-framework
azure-identity
dotenv
aiohttp